Refactor backend with the visitor pattern
