+++ chalk/core.py
@@ -5,10 +5,6 @@ import tempfile
 from dataclasses import dataclass
 from typing import Any, List, Optional, Tuple
 
-import svgwrite
-from svgwrite import Drawing
-from svgwrite.base import BaseElement
-
 import chalk.align
 import chalk.arrow
 import chalk.combinators
@@ -16,16 +12,13 @@ import chalk.model
 import chalk.types
 from chalk import transform as tx
 from chalk.envelope import Envelope
-from chalk.shape import Shape, Spacer, render_cairo_prims
+from chalk.shape import Shape
 from chalk.style import Stylable, Style
 from chalk.trace import Trace
 from chalk.transform import Affine, unit_x, unit_y
 from chalk.types import Diagram
 from chalk.utils import imgen
 
-PyCairoContext = Any
-PyLatex = Any
-PyLatexElement = Any
 Trail = Any
 Ident = Affine.identity()
 SVG_HEIGHT = 200
@@ -40,6 +33,9 @@ def set_svg_height(height: int) -> None:
 @dataclass
 class BaseDiagram(Stylable, tx.Transformable, chalk.types.Diagram):
     """Diagram class."""
+
+    def accept(self, visitor, *args, **kwargs):
+        raise NotImplementedError
 
     # Core composition
     def apply_transform(self, t: Affine) -> Diagram:  # type: ignore
@@ -113,13 +109,6 @@ class BaseDiagram(Stylable, tx.Transformable, chalk.types.Diagram):
     __floordiv__ = chalk.combinators.above2
     __add__ = chalk.combinators.atop
 
-    # Core rendering (factor out later)
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Compiles a `Diagram` to a list of `Primitive`s. The transfomation `t`
-        is accumulated upwards, from the tree's leaves.
-        """
-        raise NotImplementedError
-
     def display(
         self, height: int = 256, verbose: bool = True, **kwargs: Any
     ) -> None:
@@ -135,130 +124,14 @@ class BaseDiagram(Stylable, tx.Transformable, chalk.types.Diagram):
         # render and display the diagram
         imgen(self, **kwargs)
 
-    def render(
-        self, path: str, height: int = 128, width: Optional[int] = None
-    ) -> None:
-        """Render the diagram to a PNG file.
+    # Rendering
+    from chalk.backend.cairo import render as render_cairo
+    from chalk.backend.svg import render as render_svg
+    from chalk.backend.tikz import render as render_tikz
 
-        Args:
-            path (str): Path of the .png file.
-            height (int, optional): Height of the rendered image.
-                                    Defaults to 128.
-            width (Optional[int], optional): Width of the rendered image.
-                                             Defaults to None.
-        """
-        import cairo
-
-        envelope = self.get_envelope()
-        assert envelope is not None
-
-        pad = 0.05
-
-        # infer width to preserve aspect ratio
-        width = width or int(height * envelope.width / envelope.height)
-
-        # determine scale to fit the largest axis in the target frame size
-        if envelope.width - width <= envelope.height - height:
-            α = height / ((1 + pad) * envelope.height)
-        else:
-            α = width / ((1 + pad) * envelope.width)
-
-        surface = cairo.ImageSurface(cairo.FORMAT_ARGB32, width, height)
-        ctx = cairo.Context(surface)
-
-        s = self.scale(α).center_xy().pad(1 + pad)
-        e = s.get_envelope()
-        assert e is not None
-        s = s.translate(e(-unit_x), e(-unit_y))
-        render_cairo_prims(s, ctx, Style.root(max(width, height)))
-        surface.write_to_png(path)
-
-    def render_svg(
-        self, path: str, height: int = 128, width: Optional[int] = None
-    ) -> None:
-        """Render the diagram to an SVG file.
-
-        Args:
-            path (str): Path of the .svg file.
-            height (int, optional): Height of the rendered image.
-                                    Defaults to 128.
-            width (Optional[int], optional): Width of the rendered image.
-                                             Defaults to None.
-        """
-        pad = 0.05
-        envelope = self.get_envelope()
-
-        # infer width to preserve aspect ratio
-        assert envelope is not None
-        width = width or int(height * envelope.width / envelope.height)
-
-        # determine scale to fit the largest axis in the target frame size
-        if envelope.width - width <= envelope.height - height:
-            α = height / ((1 + pad) * envelope.height)
-        else:
-            α = width / ((1 + pad) * envelope.width)
-        dwg = svgwrite.Drawing(
-            path,
-            size=(width, height),
-        )
-
-        outer = dwg.g(
-            style="fill:white;",
-        )
-        # Arrow marker
-        marker = dwg.marker(
-            id="arrow", refX=5.0, refY=1.7, size=(5, 3.5), orient="auto"
-        )
-        marker.add(dwg.polygon([(0, 0), (5, 1.75), (0, 3.5)]))
-        dwg.defs.add(marker)
-
-        dwg.add(outer)
-        s = self.center_xy().pad(1 + pad).scale(α)
-        e = s.get_envelope()
-        assert e is not None
-        s = s.translate(e(-unit_x), e(-unit_y))
-        style = Style.root(output_size=max(height, width))
-        outer.add(s.to_svg(dwg, style))
-        dwg.save()
-
-    def render_pdf(self, path: str, height: int = 128) -> None:
-        # Hack: Convert roughly from px to pt. Assume 300 dpi.
-        heightpt = height / 4.3
-        try:
-            import pylatex
-        except ImportError:
-            print("Render PDF requires pylatex installation.")
-            return
-
-        pad = 0.05
-        envelope = self.get_envelope()
-        assert envelope is not None
-
-        # infer width to preserve aspect ratio
-        width = heightpt * (envelope.width / envelope.height)
-        # determine scale to fit the largest axis in the target frame size
-        if envelope.width - width <= envelope.height - heightpt:
-            α = heightpt / ((1 + pad) * envelope.height)
-        else:
-            α = width / ((1 + pad) * envelope.width)
-        x, _ = pad * heightpt, pad * width
-
-        # create document
-        doc = pylatex.Document(documentclass="standalone")
-        # document_options= pylatex.TikZOptions(margin=f"{{{x}pt {x}pt {y}pt {y}pt}}"))
-        # add our sample drawings
-        diagram = self.scale(α).reflect_y().pad(1 + pad)
-        envelope = diagram.get_envelope()
-        assert envelope is not None
-        padding = Primitive.from_shape(
-            Spacer(envelope.width, envelope.height)
-        ).translate(envelope.center.x, envelope.center.y)
-        diagram = diagram + padding
-        with doc.create(pylatex.TikZ()) as pic:
-            for x in diagram.to_tikz(pylatex, Style.root(max(height, width))):
-                pic.append(x)
-        doc.generate_tex(path.replace(".pdf", "") + ".tex")
-        doc.generate_pdf(path.replace(".pdf", ""), clean_tex=False)
+    render = render_cairo
+    render_svg = render_svg
+    render_pdf = render_tikz
 
     def _repr_svg_(self) -> str:
         global SVG_HEIGHT
@@ -268,14 +141,6 @@ class BaseDiagram(Stylable, tx.Transformable, chalk.types.Diagram):
         svg = open(f.name).read()
         os.unlink(f.name)
         return svg
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Convert a diagram to SVG image."""
-        raise NotImplementedError
-
-    def to_tikz(self, pylatex: PyLatex, style: Style) -> List[PyLatexElement]:
-        """Convert a diagram to SVG image."""
-        raise NotImplementedError
 
     # Getters
     def get_subdiagram_envelope(
@@ -366,49 +231,8 @@ class Primitive(BaseDiagram):
         new_transform = t * self.transform
         return self.shape.get_trace().apply_transform(new_transform)
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives.
-
-        Args:
-            t (Transform): A transform object
-                           Defaults to Ident.
-
-        Returns:
-            List[Primitive]: List of primitives.
-        """
-        return [self.apply_transform(t)]
-
-    def to_svg(self, dwg: Drawing, other_style: Style) -> BaseElement:
-        """Convert a diagram to SVG image."""
-        style = self.style.merge(other_style)
-        style_svg = style.to_svg()
-        transform = tx.to_svg(self.transform)
-        inner = self.shape.render_svg(dwg, style)
-        if not style_svg and not transform:
-            return inner
-        else:
-            if not style_svg:
-                style_svg = ";"
-            g = dwg.g(transform=transform, style=style_svg)
-            g.add(inner)
-            return g
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, other_style: Style
-    ) -> List[PyLatexElement]:
-        """Convert a diagram to SVG image."""
-
-        transform = tx.to_tikz(self.transform)
-        style = self.style.merge(other_style)
-        inner = self.shape.render_tikz(pylatex, style)
-        if not style and not transform:
-            return [inner]
-        else:
-            options = {}
-            options["cm"] = tx.to_tikz(self.transform)
-            s = pylatex.TikZScope(options=pylatex.TikZOptions(**options))
-            s.append(inner)
-            return [s]
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_primitive(self, *args, **kwargs)
 
 
 @dataclass
@@ -422,19 +246,8 @@ class Empty(BaseDiagram):
     def get_trace(self, t: Affine = Ident) -> Trace:
         return Trace.empty()
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives."""
-        return []
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Converts to SVG image."""
-        return dwg.g()
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, style: Style
-    ) -> List[PyLatexElement]:
-        """Converts to SVG image."""
-        return []
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_empty(self, *args, **kwargs)
 
 
 def empty() -> Diagram:
@@ -466,24 +279,8 @@ class Compose(BaseDiagram):
             bb = self.diagram2.get_subdiagram(name, t)
         return bb
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives."""
-        return self.diagram1.to_list(t) + self.diagram2.to_list(t)
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Converts to SVG image."""
-        g = dwg.g()
-        g.add(self.diagram1.to_svg(dwg, style))
-        g.add(self.diagram2.to_svg(dwg, style))
-        return g
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, style: Style
-    ) -> List[PyLatexElement]:
-        """Converts to tikz image."""
-        return self.diagram1.to_tikz(pylatex, style) + self.diagram2.to_tikz(
-            pylatex, style
-        )
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_compose(self, *args, **kwargs)
 
 
 @dataclass
@@ -508,28 +305,8 @@ class ApplyTransform(BaseDiagram):
         """Get the bounding envelope of the sub-diagram."""
         return self.diagram.get_subdiagram(name, t * self.transform)
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives."""
-        t_new = t * self.transform
-        return [
-            prim.apply_transform(t_new) for prim in self.diagram.to_list(t)
-        ]
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Converts to SVG image."""
-        g = dwg.g(transform=tx.to_svg(self.transform))
-        g.add(self.diagram.to_svg(dwg, style))
-        return g
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, style: Style
-    ) -> List[PyLatexElement]:
-        options = {}
-        options["cm"] = tx.to_tikz(self.transform)
-        s = pylatex.TikZScope(options=pylatex.TikZOptions(**options))
-        for x in self.diagram.to_tikz(pylatex, style):
-            s.append(x)
-        return [s]
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_apply_transform(self, *args, **kwargs)
 
 
 @dataclass
@@ -552,20 +329,8 @@ class ApplyStyle(BaseDiagram):
     ) -> Optional[Tuple[Diagram, Affine]]:
         return self.diagram.get_subdiagram(name, t)
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives."""
-        return [
-            prim.apply_style(self.style) for prim in self.diagram.to_list(t)
-        ]
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Converts to SVG image."""
-        return self.diagram.to_svg(dwg, self.style.merge(style))
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, style: Style
-    ) -> List[PyLatexElement]:
-        return self.diagram.to_tikz(pylatex, self.style.merge(style))
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_apply_style(self, *args, **kwargs)
 
 
 @dataclass
@@ -592,17 +357,5 @@ class ApplyName(BaseDiagram):
         else:
             return None
 
-    def to_list(self, t: Affine = Ident) -> List[Diagram]:
-        """Returns a list of primitives."""
-        return [prim for prim in self.diagram.to_list(t)]
-
-    def to_svg(self, dwg: Drawing, style: Style) -> BaseElement:
-        """Converts to SVG image."""
-        g = dwg.g()
-        g.add(self.diagram.to_svg(dwg, style))
-        return g
-
-    def to_tikz(
-        self, pylatex: PyLatexElement, style: Style
-    ) -> List[PyLatexElement]:
-        return self.diagram.to_tikz(pylatex, style)
+    def accept(self, visitor, *args, **kwargs):
+        return visitor.visit_apply_name(self, *args, **kwargs)
